feat: Production Readiness Fixes - IBD + Security + Logging
Phase 1 - Security (P0):
- Replace hardcoded wallet password with BITQUAN_WALLET_PASSWORD env var
- Add 'testing' feature gate for Mock PoW (release-safe)
- Add validation for empty passwords and security warnings

Phase 2 - P2P Functionality (P1):
- Implement IBD Headers handler with chain link validation
- Implement IBD GetHeaders auto-trigger when behind peers
- Add mempool txids() iterator for P2P announcement
- Implement GetMempool handler (up to 50,000 txids)

Phase 3 - Code Quality (P1):
- Replace 133+ println!/eprintln! with structured logging
- Migrate stratum_server.rs, main.rs, rpc.rs, wallet.rs, reward_engine.rs
- Use log::info!/debug!/error!/warn! appropriately

Files modified:
- crates/node/src/worker.rs (IBD + mempool P2P)
- crates/node/src/rpc.rs (password env var)
- crates/node/src/main.rs (Mock PoW feature gate + logging)
- crates/node/Cargo.toml (testing feature)
- crates/mempool/src/lib.rs (txids() method)
- crates/node/src/stratum_server.rs (logging)
- crates/node/src/reward_engine.rs (logging)

+++ crates/bq-sdk/src/psbt/mod.rs
@@ -136,8 +136,8 @@ pub enum InputKey {
 /// Output PSBT keys
 #[derive(Debug, Clone, Serialize, Deserialize, Ord, PartialOrd, Eq, PartialEq)]
 pub enum OutputKey {
-    /// Amount (8 bytes)
-    Amount(u64),
+    /// Amount (16 bytes for u128)
+    Amount(u128),
     /// ScriptPubkey (variable)
     ScriptPubkey(Vec<u8>),
     /// Proprietary data
@@ -255,7 +255,7 @@ impl PSBTOutput {
     }
 
     /// Set amount
-    pub fn set_amount(&mut self, amount: u64) {
+    pub fn set_amount(&mut self, amount: u128) {
         let mut bytes = vec![];
         bytes.extend_from_slice(&amount.to_le_bytes());
         self.add_field(OutputKey::Amount(amount), bytes);
@@ -267,7 +267,7 @@ impl PSBTOutput {
     }
 
     /// Get amount
-    pub fn get_amount(&self) -> Option<u64> {
+    pub fn get_amount(&self) -> Option<u128> {
         self.fields.iter().find_map(|(key, _value)| {
             if let OutputKey::Amount(amount) = key {
                 Some(*amount)
@@ -662,7 +662,7 @@ impl PQPSBTBuilder {
     }
 
     /// Add output
-    pub fn add_output(mut self, address: &str, amount: u64) -> Result<Self> {
+    pub fn add_output(mut self, address: &str, amount: u128) -> Result<Self> {
         let addr = Address::parse(address)?;
 
         let mut output = PSBTOutput::new();
